завершает работу с модулями
